pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]


[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
openai = "^1.33.0"
streamlit = "^1.35.0"
textblob = "^0.18.0.post0"
youtube-transcript-api = "^0.6.2"
google-api-python-client = "^2.132.0"
python-dotenv = "^1.0.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
